+++ components/style/values/computed/box.rs
@@ -13,67 +13,14 @@ use crate::values::generics::box_::{
 use crate::values::specified::box_ as specified;
 
 pub use crate::values::specified::box_::{
-    AnimationName, AnimationTimeline, Appearance, BreakBetween, BreakWithin,
-    Clear as SpecifiedClear, Contain, ContainerName, ContainerType, ContentVisibility, Display,
-    Float as SpecifiedFloat, Overflow, OverflowAnchor, OverflowClipBox, OverscrollBehavior,
-    ScrollAxis, ScrollSnapAlign, ScrollSnapAxis, ScrollSnapStop, ScrollSnapStrictness,
-    ScrollSnapType, ScrollTimelineName, ScrollbarGutter, TouchAction, TransitionProperty,
-    WillChange,
+    Appearance, BreakBetween, BreakWithin, Clear as SpecifiedClear, Contain, ContainerName,
+    ContainerType, ContentVisibility, Display, Float as SpecifiedFloat, Overflow, OverflowAnchor,
+    OverflowClipBox, OverscrollBehavior, ScrollSnapAlign, ScrollSnapAxis, ScrollSnapStop,
+    ScrollSnapStrictness, ScrollSnapType, ScrollbarGutter, TouchAction, WillChange,
 };
-
-use std::fmt::{self, Write};
-use style_traits::{ToCss, CssWriter};
 
 /// A computed value for the `vertical-align` property.
 pub type VerticalAlign = GenericVerticalAlign<LengthPercentage>;
-
-/// A computed value for the `animation-iteration-count` property.
-#[derive(Clone, Copy, Debug, MallocSizeOf, PartialEq, ToResolvedValue, ToShmem)]
-#[repr(C)]
-pub struct AnimationIterationCount(pub f32);
-
-impl ToComputedValue for specified::AnimationIterationCount {
-    type ComputedValue = AnimationIterationCount;
-
-    #[inline]
-    fn to_computed_value(&self, context: &Context) -> Self::ComputedValue {
-        AnimationIterationCount(match *self {
-            specified::AnimationIterationCount::Number(n) => n.to_computed_value(context).0,
-            specified::AnimationIterationCount::Infinite => std::f32::INFINITY,
-        })
-    }
-
-    #[inline]
-    fn from_computed_value(computed: &Self::ComputedValue) -> Self {
-        use crate::values::specified::NonNegativeNumber;
-        if computed.0.is_infinite() {
-            specified::AnimationIterationCount::Infinite
-        } else {
-            specified::AnimationIterationCount::Number(NonNegativeNumber::new(computed.0))
-        }
-    }
-}
-
-impl AnimationIterationCount {
-    /// Returns the value `1.0`.
-    #[inline]
-    pub fn one() -> Self {
-        Self(1.0)
-    }
-}
-
-impl ToCss for AnimationIterationCount {
-    fn to_css<W>(&self, dest: &mut CssWriter<W>) -> fmt::Result
-    where
-        W: Write,
-    {
-        if self.0.is_infinite() {
-            dest.write_str("infinite")
-        } else {
-            self.0.to_css(dest)
-        }
-    }
-}
 
 /// A computed value for the `contain-intrinsic-size` property.
 pub type ContainIntrinsicSize = GenericContainIntrinsicSize<NonNegativeLength>;
