+++ components/style/values/computed/ui.rs
@@ -6,7 +6,7 @@
 
 use crate::values::computed::color::Color;
 use crate::values::computed::image::Image;
-use crate::values::computed::{LengthPercentage, Number};
+use crate::values::computed::Number;
 use crate::values::generics::ui as generics;
 
 pub use crate::values::specified::ui::CursorKind;
@@ -20,6 +20,3 @@ pub type CursorImage = generics::GenericCursorImage<Image, Number>;
 
 /// A computed value for `scrollbar-color` property.
 pub type ScrollbarColor = generics::GenericScrollbarColor<Color>;
-
-/// A computed value for the `view-timeline-inset` property.
-pub type ViewTimelineInset = generics::GenericViewTimelineInset<LengthPercentage>;
